chore(homepage): extract duplicated header into a own component
